Undo most of automatic changes to CITATION.cff
The information provided by the C01 project in 6207b03 was more
complete (and, in the case of MA name spelling, more correct!) than
the information scraped from the SFB website; e.g. it included
ORCIDs. This commit brings back the original contents of the file,
with the following changes done manually:

- updated title to match the pattern used for other projects
- updated affiliation of MA
- added HW (she's listed on sfb website on project page)
- removed redundant space